Upgraded to cocos2d v1.0.0-rc3

+++ libs/cocos2d/Support/CCArray.m
@@ -1,7 +1,7 @@
 /*
  * cocos2d for iPhone: http://www.cocos2d-iphone.org
  *
- * Copyright (c) 2010 ForzeFied Studios S.L. http://forzefield.com
+ * Copyright (c) 2010 ForzeField Studios S.L. http://forzefield.com
  * 
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -224,7 +224,7 @@
 	{
 		//floor it since in case of a oneven number the number of swaps stays the same
 		int count = (int) floorf(data->num/2.f); 
-		uint maxIndex = data->num - 1;
+		NSUInteger maxIndex = data->num - 1;
 		
 		for (int i = 0; i < count ; i++)
 		{
